Remove tiles from graphics item hierarchy.
After testing showed performance issues for zooming with some documents
(and performance of the presentation view being completely useless),
profiling showed that a lot of time was spent with QGraphicsView
analyzing the item relationships between page and tile items (with or
without tiling enabled in the settings).
Hence tiling was rewritten with tile items not being graphical items anymore,
but just objects that handle rendering of some rectangular part of the page.
All painting (and visibility checking of the tiles) is now done
on the page level again which results in performance that is
comparable to previous releases without tiling support.